move CloseButton into Button
