+++ src/components/common/Button/Button.styles.ts
@@ -51,4 +51,15 @@ export default createUseStyles((theme: ITheme) => ({
       backgroundColor: theme.backgroundColorAlt,
     },
   },
+
+  CloseButton: {
+    width: "2rem",
+    lineHeight: "2rem",
+    height: "2rem",
+    padding: 0,
+    borderRadius: "50%",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+  },
 }));
