+++ src/components/common/Button/index.ts
@@ -1,3 +1,4 @@
 import Button from "./Button";
+export { CloseButton } from "./Button";
 
 export default Button;
