--- src/components/common/CloseButton/CloseButton.styles.ts
@@ -1,14 +1,0 @@
-import { createUseStyles } from "react-jss";
-
-export default createUseStyles({
-  CloseButton: {
-    width: "2rem",
-    lineHeight: "2rem",
-    height: "2rem",
-    padding: 0,
-    borderRadius: "50%",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
