--- src/components/common/CloseButton/index.ts
@@ -1,3 +1,0 @@
-import CloseButton from "./CloseButton";
-
-export default CloseButton;
